+++ src/test/java/dev/chojo/ember/feature/system/service/DemoServiceTest.java
@@ -21,7 +21,6 @@ import dev.chojo.ember.feature.board.service.BoardTicketService;
 import dev.chojo.ember.feature.board.service.FederatedBoardService;
 import dev.chojo.ember.feature.checklist.repository.ChecklistRepository;
 import dev.chojo.ember.feature.checklist.service.ChecklistService;
-import dev.chojo.ember.feature.cluster.service.ClusterService;
 import dev.chojo.ember.feature.comment.service.CommentService;
 import dev.chojo.ember.feature.events.repository.EventFederationRepository;
 import dev.chojo.ember.feature.events.repository.EventRegistrationFieldRepository;
@@ -305,10 +304,7 @@ class DemoServiceTest extends RepositoryTestBase {
                 exchangeService,
                 procurementService,
                 itemCustodyService);
-        var clusterSeeder = new DemoClusterSeeder(
-                new ClusterService(clusterRepo, stationRepo, clusterItemReleaseService, new DomainEventBus(Set.of())),
-                clusterApplicationRepo,
-                stationRepo);
+        var clusterSeeder = new DemoClusterSeeder(clusterService, clusterApplicationRepo, stationRepo);
         var formSeeder = new DemoFormSeeder(formRepo, restrictionService);
         var notificationSeeder = new DemoNotificationSeeder(
                 notificationRepo, inventoryRepo, boardService, boardTicketService, procedureService, lendingService);